backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
openai
pydantic
pytest
httpx
pytest-asyncio

